Moved Web-Endpoint to Controller class
